- Cleared out Quote Controllers (will add in later) - Initial Quote service - Initial Quote tests
Ran into some weirdness in the QuoteServiceTest; it couldn't find the QuoteService even though the service was annotated with @Component and was in a package that should have been part of the package scan. Pulling it in with a @Bean method worked, but we never figured out why the other way did not work.